dynamic array changes to drivers compile, still needs testing added lots more .cvsignore settings for eclipse

+++ server/drivers/laser/RS4LeuzeLaserDriver.cc
@@ -296,7 +296,7 @@ RS4LeuzeLaserDriver::RS4LeuzeLaserDriver(ConfigFile* cf, int section)
 	laser_ON = 1;
 
 	myLaser=new Claser(ScanPoints);
-	Data.ranges = new double[ScanPoints];
+	Data.ranges = new float[ScanPoints];
     return;
 }
