+++ server/drivers/laser/pbs_driver.cc
@@ -233,7 +233,7 @@ PBSDriver::PBSDriver(ConfigFile* cf, int section)
   data_packet.resolution = DTOR(1.75);
   data_packet.max_range = 3.5;
   data_packet.ranges_count = NUMBER_OF_RANGE_READINGS;
-  data_packet.ranges = new double [NUMBER_OF_RANGE_READINGS];
+  data_packet.ranges = new float [NUMBER_OF_RANGE_READINGS];
   data_packet.intensity_count = 0;
   data_packet.intensity = NULL;
   data_packet.id = 1;
